Reorganize rpc package
